move pillow image logic to a dedicated utils
